Add initial code for topics, add clang format
